refactor: Apply review
