+++ changes/11245.breaking.md
@@ -1,0 +1,1 @@
+Drop the valkey-based deployment live_stat path; Endpoint.live_stat and Routing.live_stat GraphQL fields now raise DeprecatedAPI and clients must migrate to the new metric query path
