--- changes/11245.enhance.md
@@ -1,1 +1,0 @@
-Drop the valkey-based deployment live_stat path; Endpoint.live_stat and Routing.live_stat GraphQL fields now raise NotImplementedAPI and clients must migrate to the new metric query path
